+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,76 @@
 
 ### Patch Changes
 
+- Updated dependencies [dee8b05]
+  - ai@6.0.0-beta.29
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [521c537]
+  - ai@5.1.0-beta.28
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.27
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [c99da05]
+  - ai@5.1.0-beta.26
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [457f1c6]
+  - ai@5.1.0-beta.25
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [90e5bdd]
+  - ai@5.1.0-beta.24
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.23
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [046aa3b]
+- Updated dependencies [e8109d3]
+- Updated dependencies [a5e152d]
+- Updated dependencies [21e20c0]
+  - ai@5.1.0-beta.22
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@5.1.0-beta.21
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [846e80e]
+  - ai@5.1.0-beta.20
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [aa0515c]
 - Updated dependencies [e7d9b00]
 - Updated dependencies [b1aeea7]